+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Abou-Auda 2024 gentamicin ([doi:10.1016/j.jsps.2024.102060](https://doi.org/10.1016/j.jsps.2024.102060)) -- children aged 1-14 years with and without acute lymphoblastic leukemia.
+
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.
 
 - Add Ravva 2010 varenicline exposure-response models ([doi:10.1038/clpt.2009.282](https://doi.org/10.1038/clpt.2009.282)) -- adult cigarette smokers in smoking-cessation trials.